+++ tests/Entity/ServiceFieldTest.php
@@ -15,7 +15,7 @@ use App\Entity\FieldAssert;
 use App\Entity\FieldOption;
 use App\Entity\ServiceField;
 use PHPUnit\Framework\TestCase;
-use \InvalidArgumentException;
+use InvalidArgumentException;
 
 /**
  * @author Christian Siewert <christian@sieware.international>